requirements/prod.txt
djangae==0.9.6
django>=1.8,<1.10
git+https://github.com/mozilla/django-csp.git#egg=djangocsp
git+https://github.com/adamalton/django-csp-reports.git#egg=cspreports
django-session-csrf
six

# Uncomment these to use mapreduce or djangae.contrib.uniquetool (which uses mapreduce)
# These are forks of Google's libraries, which contain various fixes and patches
#git+https://github.com/potatolondon/appengine-pipelines.git
#git+https://github.com/potatolondon/potato-mapreduce.git#egg=mapreduce


djangorestframework==3.2.4
django-rest-swagger==0.3.5
pytz==2015.7
django-cors-headers==1.1.0
pycrypto==2.6.1
tweepy==3.5.0
django-adminplus==0.5
jaconv==0.2.0
GoogleAppEngineCloudStorageClient==1.9.22.1
beautifulsoup4==4.5.1
lxml==3.6.4